requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
dicttoxml==1.7.4
diff-match-patch==20200713
Django==2.2.6
django-allauth==0.40.0
django-import-export==2.4.0
django-planfix==0.49
django-robokassa3==1.4
django-sitetree==1.14.0
django-tinymce4-lite==1.7.5
et-xmlfile==1.0.1
get==2019.4.13
idna==2.8
jdcal==1.4.1
jsmin==2.2.2
MarkupPy==1.14
numpy==1.16.3
oauthlib==3.0.1
odfpy==1.4.1
openpyxl==3.0.5
pandas==0.24.2
post==2019.4.13
public==2019.4.13
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.3.1
query-string==2019.4.13
request==2019.4.13
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
tablib==2.0.0
urllib3==1.24.3
xlrd==1.2.0
XlsxWriter==1.1.5
xlwt==1.3.0
xmltodict==0.12.0
